requirements.txt
Django==4.1.9
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
psycopg2
psycopg2-binary
Pillow